dependencies update and package adjust
