+++ build.gradle
@@ -27,6 +27,9 @@ repositories {
 dependencies {
     implementation('org.springframework.boot:spring-boot-starter-data-jpa')
     implementation('org.springframework.boot:spring-boot-starter-web')
+    implementation 'com.mysql:mysql-connector-j:8.0.31'
     runtimeOnly('com.h2database:h2')
+    runtimeOnly 'com.mysql:mysql-connector-j'
     testImplementation('org.springframework.boot:spring-boot-starter-test')
 }
+
