+++ src/main/java/com/kodilla/ecommercee/repository/GenericEntity.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.repository;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
