+++ src/test/java/com/kodilla/ecommercee/SpringBootJPAIntegrationTest.java
@@ -1,5 +1,7 @@
 package com.kodilla.ecommercee;
 
+import com.kodilla.ecommercee.repository.GenericEntity;
+import com.kodilla.ecommercee.repository.GenericEntityRepository;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
